refactor: utils split to front-end and back-end scripts
