+++ fairmlhealth/mimic_data.py
@@ -15,7 +15,7 @@ import numpy as np
 import os
 import pandas as pd
 import requests
-from .utils import cb_round
+from .tools import cb_round
 
 
 def load_icd_ccs_xwalk(code_type):
